requirements-test.txt
unittest2>=0.5.1
coverage>=3.5.2
Django>=1.4.1
nose>=1.2.1
setuptools-lint==0.3
pylint>=0.26.0
mock>=1.0.0
python-social-auth>=0.2.1
social-auth-app-django
wheel>=0.24.0
certifi
python-coveralls
